adding scripts - camera mode
scripting camera mode to detect animals in view. Camera overlay turns green if it detects and red if it doesn't. .. Stopping is now only allowed in slow down zones and the player can only pull out the camera when stopped. When the player captures the photo of the animal, text will appear on screen saying what animal was taken a photo of. When player resumes cart, camera mode will go away.